feat: rename ScheduledEvent to Event
I was previously thinking that I would use the term "Event" for something else,
similar to the nomenclature used by SimPy. However, after implementing
Condition, I think "Event" will be used.

The benefit with this name change is also that there is significantly less
typing when working with this library.